refactor: consolidate DNS logic into client/dns package
Move DNS cache, lookup utilities, and shared constants to client/dns/.
