+++ util/net_test.go
@@ -21,9 +21,3 @@ func TestIP(t *testing.T) {
 	assert.True(t, IsIPV6("::0"))
 	assert.False(t, IsIPV6("127.0.1"))
 }
-
-func TestLookupIPV4From(t *testing.T) {
-	ips, err := LookupIPV4From("119.29.29.29:53", "dnspod.cn")
-	assert.Nil(t, err)
-	assert.Greater(t, len(ips), 0)
-}
